requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
exceptiongroup==1.1.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gspread==5.9.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
linkedin-api==2.0.1
lxml==4.9.2
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.3
urllib3==1.26.16
wsproto==1.2.0
